+++ apps/frontend/src/pages/user/user.tsx
@@ -105,6 +105,7 @@ const User: React.FC = () => {
 						{(profileUser as UserAuthResponseDto).lastName}
 					</p>
 					<Button
+						className={styles["follow-button"]}
 						iconName={isFollowing ? "cross" : "add"}
 						label={isFollowing ? "Following" : "Follow"}
 						onClick={isFollowing ? handleUnfollow : handleFollow}